+++ lib/base/src/main/java/com/blankj/lib/base/BaseApplication.kt
@@ -6,7 +6,11 @@ import android.support.multidex.MultiDex
 import com.blankj.utilcode.util.AppUtils
 import com.blankj.utilcode.util.CrashUtils
 import com.blankj.utilcode.util.LogUtils
+<<<<<<< HEAD
 import com.blankj.utilcode.util.MessengerUtils
+=======
+import com.blankj.utilcode.util.ProcessUtils
+>>>>>>> master
 import com.squareup.leakcanary.LeakCanary
 
 /**
@@ -75,7 +79,7 @@ open class BaseApplication : Application() {
                         return "LogUtils Formatter ArrayList { " + list.toString() + " }"
                     }
                 })
-        LogUtils.file(config.toString())
+        LogUtils.i(config.toString())
     }
 
     private fun initCrash() {