+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* `19/03/08` [add] LogUtils support multi process. Publish v1.23.7.
+* `19/03/02` [fix] LogUtils#file.
 * `19/02/28` [fix] ImageUtils#calculateInSampleSize. Publish v1.23.6.
 * `19/02/26` [fix] UriUtils#uri2File. Publish v1.23.5.
 * `19/01/31` [add] HttpUtils.